Add useIsFontLoaded
