+++ src/components/QuranReader/ReadingView/Page.tsx
@@ -8,10 +8,10 @@ import Line from './Line';
 import styles from './Page.module.scss';
 import PageFooter from './PageFooter';
 
-import { selectLoadedFontFaces } from 'src/redux/slices/QuranReader/font-faces';
+import useIsFontLoaded from 'src/components/QuranReader/hooks/useIsFontLoaded';
 import { selectWordByWordPreferences } from 'src/redux/slices/QuranReader/readingPreferences';
 import QuranReaderStyles from 'src/redux/types/QuranReaderStyles';
-import { getLineWidthClassName, isQCFFont } from 'src/utils/fontFaceHelper';
+import { getLineWidthClassName } from 'src/utils/fontFaceHelper';
 import { FALLBACK_FONT, QuranFont } from 'types/QuranReader';
 import Verse from 'types/Verse';
 
@@ -35,14 +35,7 @@ const Page = ({ verses, pageNumber, quranReaderStyles, pageIndex }: PageProps) =
   const isWordByWordLayout = showWordByWordTranslation || showWordByWordTransliteration;
   const isBigTextLayout =
     isWordByWordLayout || quranTextFontScale > 3 || quranFont === QuranFont.Tajweed;
-  const loadedFonts = useSelector(selectLoadedFontFaces);
-
-  const isFontLoaded = useMemo(() => {
-    if (!isQCFFont(quranFont)) {
-      return true;
-    }
-    return loadedFonts.includes(`p${pageNumber}-${quranFont.replace('code_', '')}`);
-  }, [pageNumber, loadedFonts, quranFont]);
+  const isFontLoaded = useIsFontLoaded(pageNumber, quranFont);
 
   return (
     <div
